requirements.txt
torch
torchvision
torchaudio
flake8
isort
black
ipdb
pyannote.audio
soundfile
librosa
fastapi
uvicorn
openpyxl
pandas
